% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_weights_eval.R
\name{centroid_shift}
\alias{centroid_shift}
\title{Centroid shift: distance between weight centroid and data centroid}
\usage{
centroid_shift(w, ta_grid, pr_grid, ta_obs, pr_obs)
}
\arguments{
\item{w}{Weights}

\item{ta_grid}{Grid temperature values}

\item{pr_grid}{Grid precipitation values}

\item{ta_obs}{Observed temperature values}

\item{pr_obs}{Observed precipitation values}
}
\value{
Euclidean distance in standardized units
}
\description{
Large values suggest the surface mass is offset from where data actually is.
Computed in standardized units to make T and P comparable.
}
