% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_weights_eval.R
\name{compute_log_score}
\alias{compute_log_score}
\title{Compute log-score via KNN mapping}
\usage{
compute_log_score(w, ta_grid, pr_grid, ta_obs, pr_obs, mapping = "knn", k = 5)
}
\description{
Compute log-score via KNN mapping
}
\keyword{internal}
