% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_surface_weights_mvn.R
\name{compute_scenario_surface_weights_mvnorm}
\alias{compute_scenario_surface_weights_mvnorm}
\title{Deprecated: use compute_scenario_surface_weights_mvn()}
\usage{
compute_scenario_surface_weights_mvnorm(...)
}
\value{
See compute_scenario_surface_weights_mvn.
}
\description{
Backwards-compatible wrapper for the previous function name.
}
