% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_weights_evaluate.R
\name{compute_zero_mass_rate}
\alias{compute_zero_mass_rate}
\title{Compute zero-mass rate}
\usage{
compute_zero_mass_rate(
  w,
  ta_grid,
  pr_grid,
  ta_obs,
  pr_obs,
  mapping = "knn",
  k = 5
)
}
\description{
Compute zero-mass rate
}
\keyword{internal}
