% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_weights_eval.R
\name{data_centroid}
\alias{data_centroid}
\title{Data centroid (unweighted mean of observations)}
\usage{
data_centroid(ta_obs, pr_obs)
}
\description{
Data centroid (unweighted mean of observations)
}
\keyword{internal}
