% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_surface.R
\name{.ellipse_df_from_mu_S}
\alias{.ellipse_df_from_mu_S}
\title{Compute ellipse coordinates from mean and covariance}
\usage{
.ellipse_df_from_mu_S(mu, S, level, n = 181L)
}
\description{
Compute ellipse coordinates from mean and covariance
}
\keyword{internal}
