% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_surface.R
\name{.enforce_strict_monotonic}
\alias{.enforce_strict_monotonic}
\title{Enforce strict monotonicity on a numeric vector}
\usage{
.enforce_strict_monotonic(x, eps = 1e-09)
}
\description{
Enforce strict monotonicity on a numeric vector
}
\keyword{internal}
