% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_surface.R
\name{.kde_hdr_contours_df}
\alias{.kde_hdr_contours_df}
\title{Compute KDE HDR contours with plug-in or NRD bandwidth}
\usage{
.kde_hdr_contours_df(
  x,
  y,
  mass_levels,
  n = 120L,
  bw_method = c("auto", "plugin", "nrd"),
  bw_adjust = 1
)
}
\description{
Compute KDE HDR contours with plug-in or NRD bandwidth
}
\keyword{internal}
