% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_helpers.R
\name{.kde_is_pd_2x2}
\alias{.kde_is_pd_2x2}
\title{Median grid step for a (quasi-)regular axis}
\usage{
.kde_is_pd_2x2(M, tol = .Machine$double.eps)
}
\description{
Median grid step for a (quasi-)regular axis
}
\keyword{internal}
