% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_surface.R
\name{.make_diverging_bins}
\alias{.make_diverging_bins}
\title{Create diverging color bins centered on threshold}
\usage{
.make_diverging_bins(
  n_bins,
  rng,
  thr,
  fail_dir,
  col_failure = "#df0000",
  col_safe = "#0033FF",
  col_mid = "#FFFFFF",
  col_fail_light = "#FEE5D9",
  col_safe_light = "#D6E3FF"
)
}
\description{
Create diverging color bins centered on threshold
}
\keyword{internal}
