% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_surface.R
\name{.make_legend_labeller}
\alias{.make_legend_labeller}
\title{Create legend labeller with maximum label count}
\usage{
.make_legend_labeller(max_labels = 14L)
}
\description{
Create legend labeller with maximum label count
}
\keyword{internal}
