% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm_weights_pi.R
\name{.piw_entropy}
\alias{.piw_entropy}
\title{Entropy of a weight vector (internal)}
\usage{
.piw_entropy(w)
}
\description{
Computes Shannon entropy for a nonnegative weight vector.
}
\keyword{internal}
