% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_surface.R
\name{.pretty_bracketed}
\alias{.pretty_bracketed}
\title{Generate "nice" bracketed breaks for contours}
\usage{
.pretty_bracketed(rng, n_target)
}
\description{
Generate "nice" bracketed breaks for contours
}
\keyword{internal}
