% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_climate_surface.R
\name{.robust_mu_S}
\alias{.robust_mu_S}
\title{Compute robust mean and covariance using MCD}
\usage{
.robust_mu_S(x, y)
}
\description{
Compute robust mean and covariance using MCD
}
\keyword{internal}
