% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_probabilities.R
\name{.scenprob_dmvnorm2_chol}
\alias{.scenprob_dmvnorm2_chol}
\title{Bivariate normal density using pre-computed Cholesky factor}
\usage{
.scenprob_dmvnorm2_chol(x, mu, chol_L)
}
\description{
Bivariate normal density using pre-computed Cholesky factor
}
\keyword{internal}
