% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_probabilities.R
\name{.scenprob_normalize_vec}
\alias{.scenprob_normalize_vec}
\title{Normalize vector to sum to 1 (guarded)}
\usage{
.scenprob_normalize_vec(x)
}
\description{
Normalize vector to sum to 1 (guarded)
}
\keyword{internal}
