% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_probabilities.R
\name{.scenprob_scale_bw}
\alias{.scenprob_scale_bw}
\title{Convert bandwidths in original units to scaled units (per dimension)}
\usage{
.scenprob_scale_bw(bw_vec, params_ta, params_pr)
}
\description{
Convert bandwidths in original units to scaled units (per dimension)
}
\keyword{internal}
