% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_probabilities.R
\name{.scenprob_scale_step}
\alias{.scenprob_scale_step}
\title{Transform a step size into scaled space}
\usage{
.scenprob_scale_step(step, params)
}
\description{
Transform a step size into scaled space
}
\keyword{internal}
