% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_probabilities.R
\name{.scenprob_support_mask}
\alias{.scenprob_support_mask}
\title{Boolean mask of grid points inside `support`}
\usage{
.scenprob_support_mask(grid_ta, grid_pr, support)
}
\description{
Boolean mask of grid points inside `support`
}
\keyword{internal}
