% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_probabilities.R
\name{.scenprob_unique_colname}
\alias{.scenprob_unique_colname}
\title{Generate unique column name, checking for duplicates}
\usage{
.scenprob_unique_colname(grp, existing_names)
}
\description{
Generate unique column name, checking for duplicates
}
\keyword{internal}
