% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_probabilities.R
\name{.scenprob_validate_bw_vec}
\alias{.scenprob_validate_bw_vec}
\title{Validate a 2D bandwidth vector}
\usage{
.scenprob_validate_bw_vec(bw_vec)
}
\description{
Validate a 2D bandwidth vector
}
\keyword{internal}
