% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_probabilities.R
\name{.scenprob_wcov2}
\alias{.scenprob_wcov2}
\title{Weighted 2x2 covariance for (x,y) with weights w}
\usage{
.scenprob_wcov2(x, y, w)
}
\description{
Weighted 2x2 covariance for (x,y) with weights w
}
\keyword{internal}
