% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_surface_weights_utils.R
\name{.scenwgt_apply_scale}
\alias{.scenwgt_apply_scale}
\title{Apply scaling if enabled}
\usage{
.scenwgt_apply_scale(x, params)
}
\description{
Apply scaling if enabled
}
\keyword{internal}
