% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_weights_utils.R
\name{.scenwgt_area_weight_guard}
\alias{.scenwgt_area_weight_guard}
\title{Disable regular-grid area weighting if grid steps are invalid}
\usage{
.scenwgt_area_weight_guard(area_weight, dT, dP, verbose, tag)
}
\description{
Disable regular-grid area weighting if grid steps are invalid
}
\keyword{internal}
