% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_weights_mvn.R
\name{.scenwgt_dmvnorm2_chol}
\alias{.scenwgt_dmvnorm2_chol}
\title{Bivariate normal density using pre-computed Cholesky factor}
\usage{
.scenwgt_dmvnorm2_chol(x, mu, chol_L)
}
\description{
Bivariate normal density using pre-computed Cholesky factor
}
\keyword{internal}
