% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_surface_weights_utils.R
\name{.scenwgt_normalize_vec}
\alias{.scenwgt_normalize_vec}
\title{Normalize vector to sum to 1 (guarded)}
\usage{
.scenwgt_normalize_vec(x)
}
\description{
Normalize vector to sum to 1 (guarded)
}
\keyword{internal}
