% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_surface_weights_utils.R
\name{.scenwgt_scale_bw}
\alias{.scenwgt_scale_bw}
\title{Convert bandwidths in original units to scaled units (per dimension)}
\usage{
.scenwgt_scale_bw(bw_vec, params_ta, params_pr)
}
\description{
Convert bandwidths in original units to scaled units (per dimension)
}
\keyword{internal}
