% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_surface_weights_utils.R
\name{.scenwgt_scale_step}
\alias{.scenwgt_scale_step}
\title{Transform a step size into scaled space}
\usage{
.scenwgt_scale_step(step, params)
}
\description{
Transform a step size into scaled space
}
\keyword{internal}
