% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_weights_utils.R
\name{.scenwgt_support_mask}
\alias{.scenwgt_support_mask}
\title{Boolean mask of grid points inside `support`}
\usage{
.scenwgt_support_mask(grid_ta, grid_pr, support)
}
\description{
Boolean mask of grid points inside `support`
}
\keyword{internal}
