% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_surface_weights_kde.R
\name{.scenwgt_try_bw_plugin}
\alias{.scenwgt_try_bw_plugin}
\title{Try 2D plug-in bandwidth via ks::Hpi}
\usage{
.scenwgt_try_bw_plugin(ta, pr)
}
\description{
Try 2D plug-in bandwidth via ks::Hpi
}
\keyword{internal}
