% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_surface_weights_utils.R
\name{.scenwgt_unique_colname}
\alias{.scenwgt_unique_colname}
\title{Generate unique column name, checking for duplicates}
\usage{
.scenwgt_unique_colname(grp, existing_names)
}
\description{
Generate unique column name, checking for duplicates
}
\keyword{internal}
