% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_surface_weights_kde.R
\name{.scenwgt_validate_bw_vec}
\alias{.scenwgt_validate_bw_vec}
\title{Validate a 2D bandwidth vector}
\usage{
.scenwgt_validate_bw_vec(bw_vec)
}
\description{
Validate a 2D bandwidth vector
}
\keyword{internal}
