% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_surface_weights_cop.R
\name{.scenwgt_validate_bw_vec_1d}
\alias{.scenwgt_validate_bw_vec_1d}
\title{Validate a 2D bandwidth vector (marginal KDE bandwidths)}
\usage{
.scenwgt_validate_bw_vec_1d(bw_vec)
}
\description{
Validate a 2D bandwidth vector (marginal KDE bandwidths)
}
\keyword{internal}
