% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_surface_weights_cop.R
\name{.scenwgt_wcor}
\alias{.scenwgt_wcor}
\title{Weighted correlation (Pearson) with guards}
\usage{
.scenwgt_wcor(x, y, w)
}
\description{
Weighted correlation (Pearson) with guards
}
\keyword{internal}
