% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_weights_mvn.R
\name{.scenwgt_wcov2}
\alias{.scenwgt_wcov2}
\title{Weighted 2x2 covariance for (x,y) with weights w}
\usage{
.scenwgt_wcov2(x, y, w)
}
\description{
Weighted 2x2 covariance for (x,y) with weights w
}
\keyword{internal}
