% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm_weights_pi.R
\name{.skill_to_score}
\alias{.skill_to_score}
\title{Map skill metric to a score where higher is better (internal)}
\usage{
.skill_to_score(skill, metric)
}
\description{
Converts a skill metric to a "higher is better" score for weighting.
}
\keyword{internal}
