% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_weights_evaluate.R
\name{effective_n_hhi}
\alias{effective_n_hhi}
\title{Effective number of scenarios (HHI-based / Simpson)}
\usage{
effective_n_hhi(w)
}
\arguments{
\item{w}{Numeric vector of weights}
}
\value{
Effective N
}
\description{
Inverse Herfindahl-Hirschman Index. Less sensitive to small weights than entropy.
}
