% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_utils.R
\name{find_sequences_below}
\alias{find_sequences_below}
\title{Find Consecutive Sequences Below a Threshold}
\usage{
find_sequences_below(x, threshold, min_length = 3, min_start_index = 1)
}
\arguments{
\item{x}{Numeric vector to analyze.}

\item{threshold}{Numeric value. Elements strictly below this value are considered part of a sequence.}

\item{min_length}{Integer. Minimum number of consecutive elements below the threshold required for a sequence to be returned.}

\item{min_start_index}{Integer. Optional index threshold. Sequences starting before this index are ignored.
Default is \code{1} (i.e., all sequences considered).}
}
\value{
A list of integer vectors. Each element corresponds to the indices of one qualifying sequence in the original vector.
If no qualifying sequences are found, returns an empty list.
}
\description{
Identifies all sequences of consecutive values in a numeric vector
that are below a specified threshold and have a minimum desired length.
Optionally, only sequences starting after a given index are considered.
}
\examples{
x <- c(rep(5, 1000), 2, 1, 1, 0, 5, 2, 2, 2, 2)
find_sequences_below(x, threshold = 3, min_length = 3, min_start_index = 1000)

}
