% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kde.R
\name{plot_kde}
\alias{plot_kde}
\title{Plot Kernel Density Estimate as a heatmap}
\usage{
plot_kde(
  kde_long,
  obs_data,
  ta_col = "tavg",
  pr_col = "prcp",
  scenario_col = "scenario",
  value_col = "weight",
  xlim = NULL,
  ylim = NULL,
  bins = 20,
  raster_interpolate = FALSE,
  point_alpha = 0.8,
  fill_option = "viridis",
  fill_begin = 0,
  fill_end = 1,
  fill_alpha = 0.8,
  fill_name = "Weight"
)
}
\description{
Plot Kernel Density Estimate as a heatmap
}
