% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_weights_eval.R
\name{weight_gini}
\alias{weight_gini}
\title{Gini coefficient of weight distribution}
\usage{
weight_gini(w)
}
\arguments{
\item{w}{Numeric vector of weights}
}
\value{
Gini coefficient
}
\description{
0 = perfect equality (uniform), 1 = perfect inequality (all mass on one cell).
}
