% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_weights_eval.R
\name{weighted_centroid}
\alias{weighted_centroid}
\title{Weighted centroid of distribution}
\usage{
weighted_centroid(w, ta, pr)
}
\arguments{
\item{w}{Weights}

\item{ta}{Temperature values}

\item{pr}{Precipitation values}
}
\value{
Named vector c(ta = ..., pr = ...)
}
\description{
Weighted centroid of distribution
}
\keyword{internal}
