% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_probabilities.R
\name{.scenprob_kde_chunk_vectorized}
\alias{.scenprob_kde_chunk_vectorized}
\title{Vectorized weighted KDE over a chunk of grid points (Gaussian product kernel)}
\usage{
.scenprob_kde_chunk_vectorized(
  g_ta,
  g_pr,
  ta_use,
  pr_use,
  w_obs,
  inv_bw_ta,
  inv_bw_pr
)
}
\description{
Vectorized weighted KDE over a chunk of grid points (Gaussian product kernel)
}
\keyword{internal}
