% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_weights_cop.R
\name{.scenwgt_gaussian_copula_density}
\alias{.scenwgt_gaussian_copula_density}
\title{Gaussian copula density (2D) from Gaussian scores z1,z2 and correlation rho}
\usage{
.scenwgt_gaussian_copula_density(z1, z2, rho)
}
\description{
Uses a stable log-density formulation.
}
\keyword{internal}
