% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_surface_weights_kde.R
\name{.scenwgt_pick_bw_auto}
\alias{.scenwgt_pick_bw_auto}
\title{Auto bandwidth with optional plug-in method}
\usage{
.scenwgt_pick_bw_auto(
  ta,
  pr,
  dT,
  dP,
  k,
  alpha,
  bw_min,
  bw_max,
  bw_method = c("auto", "plugin", "nrd")
)
}
\description{
Auto bandwidth with optional plug-in method
}
\keyword{internal}
